[PHPStanRules] Rename CheckRequiredAutowireAutoconfigurePublicUsedInConfigServiceRule to CheckRequiredAutowireAutoconfigurePublicInConfigServiceRule
