+++ packages/phpstan-rules/tests/Rules/CheckRequiredAutowireAutoconfigurePublicInConfigServiceRule/Fixture/ConfigServiceMissingMethodCall.php
