+++ packages/phpstan-rules/tests/Rules/CheckRequiredAutowireAutoconfigurePublicInConfigServiceRule/Fixture/SkipConfigParameter.php
