+++ packages/phpstan-rules/tests/Rules/CheckRequiredAutowireAutoconfigurePublicInConfigServiceRule/Fixture/SkipConfigServiceHasAutowireAutoConfigurePublicMethodCall.php
